requirements.txt
numpy
pandas
websocket-client
python-binance
xgboost
scikit-learn

